Ran sbt build and uncommented delete dataset upon IT completion

+++ connector/src/test/java/com/google/cloud/spark/bigquery/v2/SparkBigQueryWriteTest.java
@@ -153,10 +153,10 @@ public class SparkBigQueryWriteTest {
 
     @AfterClass
     public static void close() throws Exception {
-        /*if (bigquery != null) {
+        if (bigquery != null) {
             bigquery.delete(DATASET, BigQuery.DatasetDeleteOption.deleteContents());
             logger.info("Deleted test dataset: " + DATASET);
-        }*/
+        }
     }
 
     @Test
